Fix key handling by getting key state over events
For some reason SDL is returning SDL_KEYDOWN/SDL_KEYUP events that
aren't occurring. This is either a serious bug or I've just not done it
right, most likely the later, though I don't recall any of my previous
uses of SDL suffering from this :/

Retrieving key state each loop is actually much closer to how we'll be
handling the custom input interface anyway however, so it is an
acceptable solution for keyboard keys also.